rest_app/pyproject.toml
[tool.poetry]
name = "hw n11"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2 = "^2.9.7"
SQLAlchemy = "^2.0.20"
uvicorn = {extras = ["standart"], version = "^0.23.2"}
fastapi = "^0.103.1"
alembic = "^1.12.0"
pydantic = {extras = ["email"], version = "^2.3.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
fastapi-mail = "^1.4.1"
redis = "^5.0"
pydantic-settings = "^2.0.3"
fastapi-redis-rate-limiter = "^1.0.1"
cloudinary = "^1.35.0"
libgravatar = "^1.0.4"


[tool.poetry.group.dev.dependencies]
Sphinx = "^7.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
